Adding Atlas web bulk application secrets,configmaps and service:
- Using a public temporary custom tomcat image with the Atlas NFS virtual user.
Once the atlas-web-bulk repo change is merged, the private container
image should be used and a secret to authenticate the cluster against the
EBI Gitlab server must be added
- Mounting the files/volumes the application needs to start up: bioentity_properties
- Creating confimap keys for the solr and jdbc credentials, and Mounting
them as volumes in the pods. The username and password are
treated as secrets and are interpolated from env variables by using the
expression ${JDBC_USERNAME} in the properties file.
- Creating the SOLR and JDBC credentials as secrets and injecting them
as env variables to the pods
- Creating a NodePort service to expose the application in the fixed port
32000.
- attaching the manager-script role to the deployer user